Add support for disabling/enabling the use of DB for verification
This lets a user disable the use of DB for verification purposes.  The new
options "--ignore-db" and "--use-db" toggle the state of this.  This sets
a UEFI variable called MokDB that makes MokManager prompt the user to approve
the setting after a reboot.

We refactor MokSBVar to MokToggleVar and set_validation to set_toggle, as
both MokDB and MokSB are really just toggle variables.