svlist: Add the new command svlist
A new command, svlist, is added to maintain the Security Version list.
Security Version is a mechanism to warn the user when loading a less
secure kernel. The distro maintainer can use svlist to inform shim to
update the list in the UEFI non-volatile storage.

There are two struct to represent the list:

The header:

typedef struct {
        uint32_t magic;
        uint16_t size;
        char signer[26];
        svnode_t nodes[0];
} __attribute__((packed)) svlist_t;

The node:

typedef struct {
        uint32_t dv;
        uint16_t sv;
        uint8_t padding[2];
} __attribute__((packed)) svnode_t;

The header stores the magic number ("SVer"), the size of the whole list,
and the signer name (or the distro name). The node contains the distro
version and the security version and is appended to the header.

To import a Security Version list into shim, it consists of 4 steps:

1. Convert the text format list into the binary format.
2. Sign the binary blob.
3. Create the related EFI variables
4. reboot the system.

svlist can do 1. and 3. while signing the binary blob can be done with
openssl or other programs that can create PKCS#7 signatures.

The text format is similar to CSV. The signer name comes first, and then
the distro version and the security version follow. Each field is
separated by a comma.

<signer>, <distro version>, <security version>, ...

For example:

SUSE Linux Enterprise, 1104, 5, 1202, 3, 1203, 1

Signed-off-by: Gary Lin <glin@suse.com>