mokutil: only check for Secure Boot support in options that need it
Mokutil always check whether the machine supports Secure Boot or not and
it exits with an error message if Secure Boot is not supported. But some
options don't need it to be supported and could be used regardless.

For example, shim could be installed on machines where SB is not enabled
or supported. Options such as set-verbosity could be useful in this case
even without Secure Boot.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>