mokutil: delete key/hash from the reverse request
This commit basically resurrects efa1e819a0 ("Remove in_pending_request()").

Sometimes the user may remove a package with a MOK , e.g. kernel module
package, and reinstall it, and then the key would be removed and enrolled
before the next boot. Since the key is already in MokListRT, the enrollment
would be skipped due to the duplicate key, and MokManager shows the
deletion option during next boot and confuses the user.

This commit neutralizes the given request from the existing reverse request
to avoid such hassles.

Signed-off-by: Gary Lin <glin@suse.com>