mokutil: handle the parsing error from "mok-variables"
[Based on the patch from Jan Kiszka]

For some reason, mokutil may fail to parse the data from the variables
in /sys/firmware/efi/mok-variables/, and build_mok_list() returns NULL.
In such case, we should continue to try variables in
/sys/firmware/efi/efivars or /sys/firmware/efi/vars.

To reflect the error, is_one_duplicate() has to return -1. Since the
"is_" functions are designed to return only 0 and 1, rename the function
as check_one_duplicate().

Ref: https://github.com/lcp/mokutil/issues/57

Signed-off-by: Gary Lin <glin@suse.com>